requirements.txt
Django==3.0.6
djangorestframework==3.11.0
Flask==1.1.1
gunicorn==20.0.4
h5py==2.9.0
Jinja2==2.10.3
numpy==1.17.2
numpydoc==0.9.1
opencv-python==4.2.0.34
openpyxl==3.0.0
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
Pillow==6.2.0
ply==3.11
qrcode==6.1
redis==3.5.0
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
setuptools
SQLAlchemy==1.3.9
sqlparse==0.3.1
torch==1.5.0
torchvision
virtualenv==16.6.1
Werkzeug==0.16.0
